Refactor to use epd-hidapi (#1)
Most of the code was refactored and cleaned up to suit my personal preferences, although the main project structure remains the same. The CSS and dashboard HTML were modified to fit better to the lower resolution of the display used. There is also a systemd unit file and timer instead of relying on cron to periodically schedule jobs. The main Python script is also structured to be run more frequently and compare the calendar events with the previous run so that new or modified events can be picked up within a short period of time. Lastly, the image capture was changed from chromium, which is relatively heavyweight, to use [cutycapt](https://cutycapt.sourceforge.net/) and [xvfb](https://www.x.org/releases/X11R7.6/doc/man/man1/Xvfb.1.xhtml). As of right now, this version of the project isn't using a battery, but rather an always-powered pi, so it's not pulling in any of the PiSugar tools.